move translation method from controllerbase to concrete controller
